Restructure project directories
Source files are now located in the same directories as corresponding header files.
Test sources are moved into separated test/ diectories under the src/ dir.
Move Catch's header file into 3rd-party/ dir.
The old WGC/ dir becomes src/.

NOTE: At this point project building does not work, neither with Scons
nor Visual Studio/MSBuild.
