+++ config/WGCServer.properties
@@ -3,7 +3,7 @@
 server.port = 80
 
 # root path of WebClient files
-client.root = ../../WGC/WebClient
+client.root = ../../src/WebClient
 
 # in future: [with logging subsystem initialized]@TODO
 # application.logger = ...
