+++ src/Server/testsMain.cpp
@@ -1,0 +1,2 @@
+#define CATCH_CONFIG_MAIN
+#include "3rd-party/catch.hpp"
