+++ src/WebClient/scripts/websocket-connection.js
@@ -1,0 +1,4 @@
+var socket = new WebSocket("ws://localhost/logon");
+
+socket.send("I want to log in!");
+
