--- tests/testsMain.cpp
@@ -1,2 +1,0 @@
-#define CATCH_CONFIG_MAIN
-#include "catch-framework/catch.hpp"
